Use custom ember-one-way-controls branch
Use a custom `ember-one-way-controls` branch that removes the dependency on `ember-get-helper` that is currently resulting in a missing module error in Ember 2.10. Until either `ember-get-helper` is fixed to ignore Ember 2.10 (it’s not required for Ember 2.x since it’s built in), or `ember-one-way-controls` removes this dependency it its project, this is required as a workaround.